Use module in project
