+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "keksobooking",
+  "type": "module",
   "version": "26.0.0",
   "private": true,
   "description": "Личный проект «{{projectTitle}}» от HTML Academy",
